pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
chainlit = "^1.3.2"
openai = "^1.56.2"
tenacity = "^9.0.0"
termcolor = "^2.5.0"
pydantic = "^2.9.2"
mysql-connector-python = "^9.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.2.2"
mypy = "^1.10.0"
pylint = "3.2.7"
ipykernel = "^6.29.4"
types-pyyaml = "^6.0.12.20240311"
no-implicit-optional = "^1.4"
types-beautifulsoup4 = "^4.12.0.20240504"
pytest-asyncio = "^0.24.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 150
output-format = "concise"
show-fixes = true
unsafe-fixes = false
preview = true
exclude = ["tutorials/", "tests/"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "B", # flake8-bugbear
    "UP", # pyupgrade,
    "I", # isort,
    "Q", # quote
    "W", # warning
]
ignore = ["B008"]

unfixable = ["B"]

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"

[tool.pylint]
disable = [
    "too-many-instance-attributes",
    "too-few-public-methods",
    "too-many-return-statements",
    "too-many-locals",
    "too-many-arguments",
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "import-outside-toplevel",
    "cyclic-import",
    "redefined-builtin",
    "broad-exception-caught",
    "broad-exception-raised",
    "super-init-not-called",
    "duplicate-code",
    "fixme",
    "logging-fstring-interpolation"
]
reports = false
max-branches = 20
output-format = "colorized"
max-line-length = 150
class-const-naming-style = "PascalCase"
generated-members = ["googleapiclient.*"]


[tool.mypy]
no_implicit_optional = false
disallow_untyped_decorators = false
ignore_missing_imports = true
check_untyped_defs = true
warn_return_any = false
warn_no_return = false
warn_unreachable = true
strict = true
disallow_subclassing_any = false
strict_optional = true
disable_error_code = ["arg-type"]
plugins = ["pydantic.mypy"]

